test-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.23.0
httpx>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
supabase==2.15.0
tenacity>=9.0.0
